Make taglib build explicitly against libstdc++.
For some reason, the travis builds of taglib were being built against libstdc++
while my local builds were being built against libc++. It also turns out that
taglib does not play nicely with libc++, which was causing it to crash. Making
it link against libstdc++ fixes those crashes. I am not sure whether libc++ or
taglib is to blame.

Also @rpath is set to resolve to @loader_path/../Frameworks, so I don't see a
good reason not to use it for taglib as well.