Change togglePlaylistButton state when newPlaylistButton is pressed.
Previously, with the playlist display closed, clicking the new playlist
button to create a new playlist would show the playlist display, but not
change the state of the playlist display button. This meant that to close
the playlist display again, the toggle playlist button needed to be pressed
twice, which was incorrect behavior.