Add app icon.
The icon needs some work, but I should probably push it before I put it
off forever.

I don't know how to get XCode to not move it into the root of the
repository, so I guess that's where it's going to be.

The .ai file should be compatible with Illustrator 5.5 or newer.