Make titlebar have OSX10.10 style if running on OSX10.10+
- Add a "fake" titlebar accessory (this is a OSX 10.10 feature) that is
  a blank view. This causes the titlebar to expand
- No need to draw a background in TitlebarViewNS if we are on OSX10.10,
  since the expanded titlebar already drew the right background for us
  (guarantees it's the same style as the OS default)
- Related to issue #49

I also tried to use a titlebar accessory instead of our hack, but I did
not found a way to position the accessory correctly. If I asked it to be
positioned bottom, it wouldn't be using the top part of the titlebar,
and if I asked for right, it wouldn't expand the titlebar and also would
only allow me to draw in the original titlebar area. Perhaps what we 
want to do isn't supported at the moment.